Delete client update

+++ public/client/src/ClientComponent.js
@@ -14,15 +14,17 @@
 "use strict";
 
 import React, { Component } from "react";
-import API from "../../common/utils/API";
-import URI from "../../common/utils/URI";
+import API from "Common/utils/API";
+import URI from "Common/utils/URI";
 import LoadingPage from "../../common/LoadingPage";
 import Auth from "../../home/src/Pages/Auth/Auth";
 import { MaintenanceHeader } from "./components/Partials/";
-import { Container, Row } from "../../common/grid";
-import { ErrorBoundary } from "../../common/error/ErrorBoundary";
+import { Input, Button } from 'Common/form';
+import { Container, Row } from "Common/grid";
+import { ErrorBoundary } from "Common/error/ErrorBoundary";
 import MaintenanceWrapper from "./components/Maintenance/MaintenanceWrapper";
 
+import Modal from "react-responsive-modal";
 import "react-tabs/style/react-tabs.css";
 import "./App.css";
 import "../../common/styles/animate.css";
@@ -42,9 +44,12 @@ export default class ClientComponent extends Component {
     super(props);
 
     this.state = {
-      client: null, //  should be passed in the response from Authorization
-      pageData: null, //  page data object
-      hasErrors: false //  flags if we should display a message about errors
+      client: null,       //  should be passed in the response from Authorization
+      pageData: null,     //  page data object
+      hasErrors: false,   //  flags if we should display a message about errors
+      launchModal: false,
+      deleteName: "",
+      errorDelete: "",
     };
   }
 
@@ -61,6 +66,11 @@ export default class ClientComponent extends Component {
         })
         .then(({ json, resolve }) => resolve(json));
     });
+  }
+
+  handleInputChange = event => {
+    const { name, value } = event.target;
+    this.setState({ [name]: value });
   }
 
   /**
@@ -99,13 +109,18 @@ export default class ClientComponent extends Component {
   /**
    * @todo run some security checks to make sure this is what the user wants to do.
    */
-  deletePlatform = event => {
+  deleteClient = event => {
     event.preventDefault();
-    API.client.deleteClient(this.state.client).then(resp => {
-      this.setState({ client: resp.data });
-    });
+    if(this.state.deleteName !== this.state.client.name) {
+      this.setState({ errorDelete: "Names do not match."});
+      return;
+    } else {
+      API.client.deleteClient(this.state.client).then(resp => {
+        //this.setState({ client: resp.data });
+        URI.redirect('/dashboard');
+      });
+    } 
   };
-
   /**
    * @function toggleSystem
    * @description allows a user to make their system live
@@ -121,36 +136,56 @@ export default class ClientComponent extends Component {
     this.setState({ client: client });
   };
 
+  toggleModal = () => {
+    const { launchModal } = this.state;
+    this.setState({ launchModal: !launchModal, errorDelete: "" })
+  }
 
   render() {
     return (
       <div className="app-container animated fadeIn">
-        <div>
-          {this.state.pageData === null && !this.state.redirectToLogin ? (
-            <LoadingPage />
-          ) : (
-            <div className="animated fadeIn">
-              <header className="app-header">
-                <h1 className="header-title">
-                  {this.state.pageData.main.title}
-                </h1>
-              </header>
-              {this.state.client.isDeleted ? (
-                <h2>This platform has been deleted</h2>
-              ) : (
-                <main className="app-content">
-                  <ErrorBoundary>
-                    <Container>
-                      <Row>
-                        <a href="/dashboard">
-                          <i className="fa fa-arrow-left" /> Return to Dashboard
-                        </a>
+        {this.state.pageData === null && !this.state.redirectToLogin ? (
+          <LoadingPage />
+        ) : (
+          <div className="animated fadeIn">
+            <Modal open={this.state.launchModal} className="modal" onClose={this.toggleModal } center>
+              <h2>DANGER ZONE</h2>
+              <p>You are about to remove your Network<br />
+              This action <strong>cannot</strong> be undone!</p>
+              <p><strong>Type in the name of your network exactly as it appears</strong></p>
+              <h3 className="danger-text">{ this.state.client.name }</h3>
+              <form onSubmit={ this.deleteClient } method="GET" className="floating-form">
+                <Input name="deleteName" value={ this.state.deleteName }
+                  onChange={this.handleInputChange }
+                  errorText={ this.state.errorDelete }
+                />
+                <Button style="danger" type="submit" text="Continue and Delete" />
+              </form>
+                
+              
+            </Modal>
+            <header className="app-header">
+              <h1 className="header-title">
+                {this.state.pageData.main.title}
+              </h1>
+            </header>
+            <main className="app-content">
+              <ErrorBoundary>
+                <Container>
+                  <Row>
+                    <a href="/dashboard">
+                      <i className="fa fa-arrow-left" /> Return to Dashboard
+                    </a>
+                    {this.state.client.isDeleted ? (
+                      <h2>This platform has been deleted</h2>
+                    ) : (
+                      <div className="drop-row">
                         <MaintenanceHeader
                           toggleSystem={this.toggleSystem}
                           client={this.state.client}
                           hasErrors={this.state.hasErrors}
                           headerText={this.state.pageData.header}
-                          deletePlatform={this.deletePlatform}
+                          launchModal={this.toggleModal }
                         />
                         {this.state.isSaved ? (
                           <span className="badge badge-success">
@@ -161,16 +196,15 @@ export default class ClientComponent extends Component {
                           client={this.state.client}
                           pageData={this.state.pageData}
                         />
-                      </Row>
-                    </Container>
-                  </ErrorBoundary>
-                </main>
-                /* END isDeleted CHECK */
-              )}
-            </div>
-            /* END pageData CHECK */
-          )}
-        </div>
+                      </div>
+                    )}
+                  </Row>
+                </Container>
+              </ErrorBoundary>
+            </main>
+          </div>
+          /* END pageData CHECK */
+        )}
       </div>
     );
   }
